ADVISOR-2384 (#944)
When pressing the 'Reset Filter' button on a pathways details page, the filter would reset the pathway to null, and then it would just show all recommendations regardless if they were related to the pathway.

To see changes,
Go to advisor, navigate to pathways table , and select a pathway. The default filters are currently 'systems impacted' and ' status enabled'. Before the change, if you pressed reset filter, it would show all recommendations, now it should still show recommendations for specific pathway.
This change also removes the 'pathway' param in the URL as to not
duplicate it

Co-authored-by: Adonis <adonis@localhost-live.attlocal.net>